fixup! fixup! fixup! fixup! fixup! Various U/X improvements for `helmfile apply`
